Fixed style and geolocation

+++ shared/src/main/scala/ch/wsl/box/shared/utils/JSONUtils.scala
@@ -24,7 +24,7 @@ object JSONUtils extends Logging {
     Try {
       val json:Json = typ match {
         case JSONFieldTypes.NUMBER => value.toDouble.asJson
-        case JSONFieldTypes.INTEGER => value.toLong.asJson
+        case JSONFieldTypes.INTEGER => value.toDouble.toLong.asJson
         case JSONFieldTypes.BOOLEAN => Json.fromBoolean(value.toBoolean)
         case JSONFieldTypes.JSON => parser.parse(value) match {
           case Left(value) => throw new Exception(value.message)
